requirements.txt
SpeechRecognition
PyAudio
PyAutoGUI
pygame
pydub
requests
openai
gTTS
piper-tts
python-dotenv